CaptureStill Note
Realised that sometimes actTakePicture doesn't return the image url in its response because Sony is dumb and decided to code their API that way. So an empty result doesn't actually mean the image capture failed - functionally means the app works fine, but sometimes thinks a capture failed even though it didn't. Could fix this later if it's too annoying.